datatype has to be long
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@23506 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/libutils/utils.c
@@ -437,7 +437,7 @@ char *get_tcfmark(uint32 mark)
 #endif
 
 #ifdef HAVE_AQOS
-void add_client_classes(unsigned int base, unsigned int uprate, unsigned int downrate, unsigned int lanrate, unsigned int level)
+void add_client_classes(unsigned int base, unsigned int uprate, unsigned int downrate, unsigned long lanrate, unsigned int level)
 {
 	char *wan_dev = get_wan_face();
